+++ arcore-location-extension/src/test/kotlin/org/neidhardt/arlocation/LocationArNodeTest.kt
@@ -39,4 +39,5 @@ class LocationArNodeTest {
 				0.1
 		)
 	}
+
 }